+++ src/base/build-type.ts
@@ -1,0 +1,4 @@
+export type BuildType =
+    "Series2" |
+    "Series3" |
+    "Series4"